Extract Config.parseConfigFile, move HUBOT_* vars
This is another step in decoupling the core library from Hubot-specific
details and switching to the Slack Events API.
