+++ lib/config.js
@@ -1,14 +1,10 @@
 'use strict';
 
 var fs = require('fs');
-var path = require('path');
 
 module.exports = Config;
 
-function Config(configuration, logger) {
-  var config = configuration ||
-        parseConfigFromEnvironmentVariablePathOrUseDefault(logger);
-
+function Config(config) {
   validate(config);
 
   for (var fieldName in config) {
@@ -42,6 +38,21 @@ var schema = {
   }
 };
 
+Config.parseConfigFile = function(configPath, logger) {
+  var errorPrefix = 'failed to load configuration from ' + configPath + ': ';
+  logger.info(null, 'reading configuration from', configPath);
+
+  try {
+    return JSON.parse(fs.readFileSync(configPath, 'utf8'));
+  } catch (err) {
+    if (err instanceof SyntaxError) {
+      errorPrefix = errorPrefix + 'invalid JSON: ';
+    }
+    err.message = errorPrefix + err.message;
+    throw err;
+  }
+};
+
 function validate(config) {
   var errors = [],
       errMsg;
@@ -62,32 +73,6 @@ function validate(config) {
     errMsg = 'Invalid configuration:\n  ' + errors.join('\n  ');
     throw new Error(errMsg);
   }
-}
-
-function parseConfigFromEnvironmentVariablePathOrUseDefault(logger) {
-  var configPath = (process.env.HUBOT_SLACK_GITHUB_ISSUES_CONFIG_PATH ||
-        path.join('config', 'slack-github-issues.json')),
-      errorPrefix = 'failed to load configuration from ' + configPath + ': ',
-      config;
-  logger.info(null, 'reading configuration from', configPath);
-
-  try {
-    config = JSON.parse(fs.readFileSync(configPath, 'utf8'));
-  } catch (err) {
-    if (err instanceof SyntaxError) {
-      errorPrefix = errorPrefix + 'invalid JSON: ';
-    }
-    err.message = errorPrefix + err.message;
-    throw err;
-  }
-
-  if (process.env.HUBOT_GITHUB_TOKEN) {
-    config.githubApiToken = process.env.HUBOT_GITHUB_TOKEN;
-  }
-  if (process.env.HUBOT_SLACK_TOKEN) {
-    config.slackApiToken = process.env.HUBOT_SLACK_TOKEN;
-  }
-  return config;
 }
 
 function checkRequiredTopLevelFields(config, errors) {
